% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{blum-class}
\alias{blum}
\alias{blum-class}
\title{blum class}
\description{
The class for bead level information for a Luminex xMap experiment. It
contains the fluorescence intensity for each bead as well as design
information regarding the experiment.
}
\section{Slots}{

\describe{
\item{\code{phenoData}}{An \code{AnnotatedDataFrame}. Contains the information
regarding the samples (e.g: sample_type, sample_name, well, filename, ...).}

\item{\code{featureData}}{An \code{AnnotatedDataFrame}. Contains the information
regarding the analytes: ID and name.}

\item{\code{exprs}}{A \code{data.table}. Contains the intensities measured for each
 bead.}
}}
\author{
Renan Sauteraud
}
\seealso{
\code{read.experiment}, \code{\link{ExpressionSet}}
}

