% Generated by roxygen2 (4.0.0): do not edit by hand
\name{concentration}
\alias{concentration}
\alias{concentration,slum-method}
\title{Concentration accessor}
\usage{
concentration(object)
}
\arguments{
\item{object}{A \code{slum} object.}
}
\description{
Concentration matrix accessor for slum objects.
}
\details{
The function access the concentration matrix in the \code{assayData} slot of
a \code{slum} object
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{slum}}
}

