% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot_layout}
\alias{plot_layout}
\alias{plot_layout,blumORslum-method}
\title{Plot a plate layout}
\usage{
plot_layout(object, plate_name = NULL, ...)
}
\arguments{
\item{object}{A \code{blum} or \code{slum}. The object containing the
experiment to plot.}

\item{plate_name}{A \code{character}. The name of the plate to plot.}

\item{...}{Additional arguments to be passed to ggplot2's \code{geom_polygon}
 function.}
}
\description{
Visualisation of a plate's sample. This draws a plot of a selected plate
where each graphical element can be changed to display sample information.
}
\details{
The information passed in the additional aguments must be available in the
\code{phenoData} slot of the object.
}
\examples{
#plot_layout(blum, "plate1", fill="sample_type", col = "concentration")
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{geom_polygon}}
}

