% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot_SC}
\alias{plot_SC}
\title{Plot Standard curves}
\usage{
plot_SC(slum)
}
\arguments{
\item{slum}{A \code{slum} object. The experiment to plot the standard curves.}
}
\description{
Wrapper around the geom_sc method for people who don't want to melt and
ggplot.
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{geom_sc}}
}

