% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read.experiment}
\alias{read.experiment}
\title{Read xMap experiments}
\usage{
read.experiment(path = "./")
}
\arguments{
\item{path}{A \code{character}, the directory containing the data and
mapping files.}
}
\value{
An object of class \code{blum} that is the base structure for this
package.
}
\description{
Creates a blum object from a path storing a Luminex xMap experiment.
}
\details{
The folder passed in path argument must be structured in a specific
way and  contain mapping files describing the experiment. See the LumiR user
guide for more information.
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{blum}}
}

