% Generated by roxygen2 (4.0.0): do not edit by hand
\name{set_center}
\alias{set_center}
\alias{set_center,blum,character-method}
\alias{set_center,slum,character-method}
\title{set_center}
\usage{
set_center(object, center_name)
}
\arguments{
\item{object}{A \code{blum} or \code{slum} object.}

\item{center_name}{A \code{character}. The name of the center or lab that
 generated the data in the object.}
}
\description{
Add information regarding which center or lab produced the data. This is
useful when multiple source ran the same experiment.
}
\author{
Renan Sauteraud
}
\seealso{
\code{blum}, \code{slum}
}

