% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setup_templates}
\alias{setup_templates}
\title{Create templates}
\usage{
setup_templates(path, templates = c("layout", "analyte", "phenotype"),
  write = TRUE)
}
\arguments{
\item{path}{A \code{character}. The pathname of an experiment folder. This is
the same path as the one that will be used in \code{read.experiment}.}

\item{templates}{A \code{character} vector. Can be any combination of
"layout", "analyte" and "phenotype". A mapping file will be created for each
selected template.}

\item{write}{A \code{logical}. If TRUE, write the template files in the
 given \code{path}. When \code{setup_templates} is called through
 \code{read.experiment}, write is set to FALSE and the template files are not
 saved.}
}
\value{
An invisible \code{list} of \code{data.frame} to be used internally by
\code{read.experiment} if one or more mapping files are missing from the given
path.
}
\description{
Create files to be used as mapping files. The function will guess as much
information as possible based on the data files. The created files may be
incomplete and require user modifications but they ensure that the required
columns are present.
}
\details{
If a mapping file specified by the templates list already exists, the function
will not overwrite the existing file. The files should be removed before
running the function in order to create new mapping files.
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{read.experiment}}
}

