% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{slum-class}
\alias{slum}
\alias{slum-class}
\title{slum class}
\description{
The class contains the summarized information for a Luminex xMap experiment.
It contains the median fluorescence intensity for each well and analytes as
well as information regarding the desing of the experiment.
}
\section{Slots}{

\describe{
\item{\code{phenoData}}{A \code{AnnotatedDataFrame}.Contains the information
regarding the samples (e.g: sample_type, sample_name, well, filename, ...).}

\item{\code{featureData}}{An \code{AnnotatedDataFrame}. Contains the information
regarding the analytes: ID and name.}

\item{\code{assayData}}{An \code{environment} with two identically dimensioned
matrices with analytes as rownames and sample_id as colnames.
\itemize{
  \item \code{exprs}: The \code{matrix} of the MFI.
  \item \code{concentration}: The \code{matrix} of the concentrations
    calculated for the MFIs in \code{exprs}.
}}

\item{\code{fit}}{A \code{data.frame}. Contains the information regarding the
standard curve fitting: The values of the parameters to use in \code{inv}
for each couple analyte / sample_id.}

\item{\code{formula}}{A \code{formula}. The formula used for the standard curve
fitting. By default a 5-Parameters Logistic.}

\item{\code{inv}}{A \code{function}. The inverse function of the formula.}

\item{\code{unit}}{A \code{character}, "MFI".}

\item{\code{protocolData}}{An optional \code{AnnotatedDataFrame}, inherited from
\code{ExpressionSet}.}

\item{\code{experimentData}}{An optional \code{MIAME}, inherited from
 \code{ExpressionSet}.}
}}
\author{
Renan Sauteraud
}
\seealso{
\code{slummarize}, \code{\link{ExpressionSet}}, \code{blum}
}

