% Generated by roxygen2 (4.0.0): do not edit by hand
\name{slummarize}
\alias{slummarize}
\title{Summarize a blum object}
\usage{
slummarize(from, type = "MFI")
}
\arguments{
\item{from}{A \code{blum} object. The object to summarize. The object must
have information about the standards location and concentration in its
\code{phenodata} slot. See the user guide for more information on the
requirements.}

\item{type}{A \code{character}. "MFI"}
}
\value{
An object of class \code{slum}. With two expression matrices for MFI
and concentrations.
}
\description{
Summarize a blum object into a slum object. Fit the standard curves and
calculate median fluorescence intensities (MFI) and the corresponding
concentrations for each well.
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{blum}}, \code{\link{slum}}
}

