% Generated by roxygen2 (4.0.0): do not edit by hand
\name{writeMBAA}
\alias{writeMBAA}
\title{Write a slum object as ImmPort MBAA results}
\usage{
writeMBAA(object, outfile = "./MBAA_results", type = "csv",
  concentration_unit = "pg/mL")
}
\arguments{
\item{object}{A \code{slum} object. The summarized experiment to report.}

\item{outfile}{A \code{character}. The name of the output file without extension.}

\item{type}{A \code{character}. Can be "csv" or "xls". Determines the type of
file to be written. This will also be appended to the filename given in outfile.}

\item{concentration_unit}{A \code{character}. The unit of the concentration
to be reported in the document.}
}
\description{
Format a \code{slum} object into a table as specified by ImmPort template:
'https://immport.niaid.nih.gov/example_submission_packages/MBAA_Results.xls'.
}
\author{
Renan Sauteraud
}
\seealso{
\code{\link{slum}}
}

